refactor: AfternoteSaveState 패키지 이동 및 검증 로직 개선
- `AfternoteSaveState.kt` 파일 위치를 `model` 패키지에서 `state` 패키지로 이동
- `AfternoteValidationException` 발생 시 원인 파악을 위한 에러 메시지 추가
- `AfternoteValidationError` 및 `AfternoteSaveState` 코드 내 KDoc 한글화 및 가이드 주석 보강
- `AfternoteValidationError` 생성자 파라미터의 어노테이션 위치 수정 및 코드 정리
